Template Analyzer Agent/requirements.txt
fastapi==0.128.6
uvicorn==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg>=0.31.0
python-dotenv==1.0.1
google-adk==0.1.1
google-genai
pydantic-settings==2.7.0
pydantic==2.12.5
deprecated
google-cloud-documentai
google-cloud-storage
python-multipart
pypdf
PyJWT>=2.0.0